Alumnos validaciones

+++ controllers/grados.js
@@ -1,5 +1,6 @@
 const Grado = require('../models/grados')
 const Profesor = require('../models/profesor')
+const Alumnogrado = require('../models/alumnogrado')
 const { validationResult } = require('express-validator')
 
 
@@ -24,8 +25,15 @@ const getGrados = async(req, res) => {
 }
 
 const getGrado = async(req, res) => {
-    const grado = await Grado.findById(req.params.id);
-    res.json({
+    const id = req.params.id
+    if (id.toString().length !== 24) {
+        return res.json({
+            ok: false,
+            msg: 'El id no es valido',
+        })
+    }
+    const grado = await Grado.findById(id);
+    return res.json({
         grado
     })
 }
@@ -50,6 +58,12 @@ const postGrado = async(req, res) => {
 const putGrado = async(req, res) => {
     const id = req.params.id;
     try {
+        if (id.toString().length !== 24) {
+            return res.json({
+                ok: false,
+                msg: 'El id no es valido',
+            })
+        }
         const grado = await Grado.findById(id);
         if (!grado) {
             return res.status(404).json({
@@ -70,19 +84,48 @@ const putGrado = async(req, res) => {
 const delGrado = async(req, res) => {
     const id = req.params.id;
     try {
+        if (id.toString().length !== 24) {
+            return res.json({
+                ok: false,
+                msg: 'El id no es valido',
+            })
+        }
         const grado = await Grado.findById(id);
         if (!grado) {
-            return res.status(404).json({
-                mensaje: 'Grado no encontrado'
+            return res.json({
+                ok: true,
+                msg: 'Grado no encontrado',
             })
         }
 
-        const gradoActualizado = await Grado.findByIdAndDelete(id);
-        res.json({
-            ok: true,
-            msg: 'Grado borrado',
-            res: gradoActualizado
-        })
+        ///
+        const alumnosgrados = await Alumnogrado.find();
+        if (!alumnosgrados) {
+            const gradoActualizado = await Grado.findByIdAndDelete(id);
+            return res.json({
+                ok: true,
+                msg: 'Grado borrado',
+                res: gradoActualizado
+            })
+        } else {
+            for (const prop in alumnosgrados) {
+                if (alumnosgrados[prop].GradoId === id) {
+                    return res.json({
+                        ok: false,
+                        msg: 'No se puede borrar el grado porque pertenece a un Alumnogrado, elimine el Alumnogrado primero',
+                    })
+                }
+            }
+            const gradoActualizado = await Grado.findByIdAndDelete(id);
+            res.json({
+                ok: true,
+                msg: 'Grado borrado',
+                res: gradoActualizado
+            })
+        }
+        ///
+
+
     } catch (error) {
         console.log(error);
     }
